renamed dir from diagnostics to DiagnosticSummary
