+++ packages/suite-base/src/panels/DiagnosticStatus/index.tsx
@@ -5,6 +5,7 @@ import Panel from "@lichtblick/suite-base/components/Panel";
 import DiagnosticStatusPanel from "@lichtblick/suite-base/panels/DiagnosticStatus/DiagnosticStatusPanel";
 import { DEFAULT_CONFIG } from "@lichtblick/suite-base/panels/DiagnosticStatus/constants";
 
+// Diagnostic - Detail
 export default Panel(
   Object.assign(DiagnosticStatusPanel, {
     panelType: "DiagnosticStatusPanel",
