+++ packages/suite-base/src/panels/DiagnosticSummary/DiagnosticSummary.style.ts
