+++ packages/suite-base/src/panels/DiagnosticSummary/assets/diagnostic-summary.png
