+++ packages/suite-base/src/panels/DiagnosticSummary/hooks/useDiagnostics.ts
@@ -18,7 +18,7 @@ import { useMemo } from "react";
 
 import { useMessageReducer } from "@lichtblick/suite-base/PanelAPI";
 import { DiagnosticStatusArrayMsg } from "@lichtblick/suite-base/panels/DiagnosticStatus/types";
-import { DiagnosticsById } from "@lichtblick/suite-base/panels/diagnostics/types";
+import { DiagnosticsById } from "@lichtblick/suite-base/panels/DiagnosticSummary/types";
 import { MessageEvent } from "@lichtblick/suite-base/players/types";
 
 import { computeDiagnosticInfo } from "../utils/util";
