+++ packages/suite-base/src/panels/DiagnosticSummary/hooks/useStaleTime.ts
