+++ packages/suite-base/src/panels/DiagnosticSummary/types.ts
