+++ packages/suite-base/src/panels/DiagnosticSummary/utils/settings.ts
@@ -6,8 +6,8 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import { SettingsTreeNodes } from "@lichtblick/suite";
-import { DEFAULT_SECONDS_UNTIL_STALE } from "@lichtblick/suite-base/panels/diagnostics/constants";
-import { DiagnosticSummaryConfig } from "@lichtblick/suite-base/panels/diagnostics/types";
+import { DEFAULT_SECONDS_UNTIL_STALE } from "@lichtblick/suite-base/panels/DiagnosticSummary/constants";
+import { DiagnosticSummaryConfig } from "@lichtblick/suite-base/panels/DiagnosticSummary/types";
 
 export function buildSummarySettingsTree(
   config: DiagnosticSummaryConfig,
